fix: decouple normal and gda compatibility
this is needed, as the tests can't import "Shell". But the test import db.ts > compatibility.ts > Shell, so we split the compatibility into two layers, to make the tests work again
