+++ src/utils/db.ts
@@ -1,5 +1,5 @@
 import Gda5 from '@girs/gda-5.0';
-import { add_expr_value, type DataModelIter, type SqlBuilder, unescape_string } from '@pano/utils/compatibility';
+import { add_expr_value, type DataModelIter, type SqlBuilder, unescape_string } from '@pano/utils/gda_compatibility';
 import { logger } from '@pano/utils/shell';
 
 const debug = logger('database');
